refactor: Move circom witness generation to another module
